refactor: refactored interfaces for compatibility (#608)
* refactor: refactored interfaces for compatibility

* style(examples/README): lint with pre-commit
